Add check for visibility for github_repository
Add a check for the visiblity property of the github_repository
resource from the GitHub provider. It checks if the visibility was not
set to "private". This complies with the latest version of the provider
(3.0.0+), which switches to visibility in favor of the private property.
Left the private property in for backward compatibility.

Resolves #326